Create new RLP for modded content
